# Perhaps faster normalization: Never visit a descendant twice in normalizaton