I see @czechdave  say that normalization potentially slows down the Value.fromJSON; And I think perhaps it is a result from `normalizeKeys.includes`, a O(childNodes^2) comparison.  

This PR take advantage of cache and validate to ensure speed up normalization.

I am not confident about the benchmark running on my laptop.  Could anyone help me rerun the benchmark in their own computer?
```
benchmarks
    html-serializer
      deserialize
        8.98 → 11.02 ops/sec
      serialize
        1057.71 → 1185.42 ops/sec
    plain-serializer
      deserialize
        553.52 → 536.36 ops/sec
      serialize
        9107.18 → 9525.40 ops/sec
    rendering
      normal
        111.93 → 126.17 ops/sec
    changes
      delete-backward
        105548.40 → 144896.70 ops/sec (37% faster)
      delete-forward
        19318.61 → 30794.16 ops/sec (59% faster)
      insert-text-by-key-multiple
        85.57 → 97.89 ops/sec
      insert-text-by-key
        1045.61 → 1251.19 ops/sec
      insert-text
        1071.37 → 1150.59 ops/sec
      normalize
        1435.06 → 215908.18 ops/sec (14945% faster) 😱
      split-block
        22.43 → 83.04 ops/sec (270% faster) 🙌
    models
      from-json
        126.48 → 221.79 ops/sec (75% faster)
      get-blocks-at-range
        515597.57 → 691750.84 ops/sec (34% faster)
      get-blocks
        2389340.86 → 2876387.07 ops/sec
      get-characters-at-range
        675896.57 → 780662.25 ops/sec
      get-characters
        2906877.93 → 3160090.06 ops/sec
      get-inlines-at-range
        715855.63 → 848482.59 ops/sec
      get-inlines
        2821376.68 → 3106357.04 ops/sec
      get-leaves
        1591780.52 → 1837992.19 ops/sec
      get-marks-at-range
        800173.93 → 890650.77 ops/sec
      get-marks
        3167696.93 → 3177520.59 ops/sec
      get-path
        975600.69 → 910776.69 ops/sec
      get-texts-at-range
        937712.33 → 924806.06 ops/sec
      get-texts
        3334330.56 → 3305250.49 ops/sec
      has-node-multiple
        128830.81 → 123509.47 ops/sec
      has-node
        985303.14 → 930603.71 ops/sec
      to-json
        3584.00 → 3341.62 ops/sec
      update-node
        18095.09 → 14661.38 ops/sec
```