@ianstormtaylor 2018-02-21
@zhujinxuan this is incredibly interesting! Good find. I'll pull down this locally and see if I get similar results.

Could you rename `getFirstValidationKey` to `getFirstInvalidDescendantKey`? I think that will be more clear as to what it's doing in the future for people.