@zhujinxuan 2018-02-21
@ianstormtaylor Fixed about the var name~

I am not sure whether it is resulted from the cache; that is my guess only.  I think, for such a huge improvement, the most intuitive thing is the cache.