@julienp 2018-02-22
This fixes the memory issues I encountered with large documents. I was under the impression memoization was at fault (#1402), since I was able to load these documents when I disabled memoization altogether. Pure speculation, but I wonder if browsers over allocate the `normalizedKeys` array, since it doesn't have a predictable size and is pushed into.