@zhujinxuan 2018-02-22
@julienp I am quite surprising about the memory problem by `normalizedKeys`.  Because they are just short strings, at the worse complexity O(number of all nodes in document)...