@brekk 2018-02-22
Not seeing quite the same performance change in all the same places on my machine, but `fromJSON` is 93% faster. 🙌 

<details>
```
  benchmarks
    html-serializer
      deserialize
        9.78 → 10.27 ops/sec
      serialize
        1010.85 → 1047.72 ops/sec
    plain-serializer
      deserialize
        626.74 → 725.75 ops/sec
      serialize
        7730.22 → 6872.48 ops/sec
    rendering
      normal
        86.04 → 86.98 ops/sec
    changes
      delete-backward
        157591.67 → 140251.25 ops/sec
      delete-forward
        46236.65 → 41172.02 ops/sec
      insert-text-by-key-multiple
        105.61 → 103.83 ops/sec
      insert-text-by-key
        1115.69 → 1065.21 ops/sec
      insert-text
        998.34 → 1011.77 ops/sec
      normalize
        104465.80 → 108040.58 ops/sec
      split-block
        64.76 → 93.13 ops/sec (44% faster)
    models
      from-json
        133.88 → 257.84 ops/sec (93% faster)
      get-blocks-at-range
        495634.45 → 715473.32 ops/sec (44% faster)
      get-blocks
        2630222.82 → 3282544.33 ops/sec
      get-characters-at-range
        475838.66 → 715998.18 ops/sec (50% faster)
      get-characters
        2431012.73 → 2934843.22 ops/sec
      get-inlines-at-range
        504310.26 → 728823.82 ops/sec (45% faster)
      get-inlines
        2371000.69 → 2971701.62 ops/sec
      get-leaves
        1066755.65 → 1510276.93 ops/sec (42% faster)
      get-marks-at-range
        586453.22 → 753844.17 ops/sec
      get-marks
        2441511.81 → 2850438.23 ops/sec
      get-path
        898561.00 → 761193.82 ops/sec
      get-texts-at-range
        795416.85 → 721657.42 ops/sec
      get-texts
        3327712.46 → 3013297.68 ops/sec
      has-node-multiple
        116193.33 → 89375.36 ops/sec
      has-node
        867552.13 → 796880.86 ops/sec
      to-json
        2324.30 → 1891.77 ops/sec
      update-node
        12770.59 → 14989.13 ops/sec

Done in 53.95s.
```
</details>

Running on:
OSX High Sierra 10.13.3
3.5 GHz processor
16 GB memory