@czechdave 2018-02-22
@brekk the benchmarks are quite unreliable, read up here: https://github.com/ianstormtaylor/slate/issues/1662