@zhujinxuan 2018-02-22
@ianstormtaylor Fixed about `getFirstDescendant`. Because `getFirstDescendant` can get the node directly, then I make some fix in the normalizeNodeAndChildren loop to go directly to the invalid-descendant.