@zhujinxuan 2018-02-22
@julienp According to @ianstormtaylor 's suggestion, I changed the `getFirstInvalidDescendant`'s return to the node value;  Can you have a check about the new cache result?