@zhujinxuan 2018-02-22
New benchmark result after the last commit fix `getFirstInvidDescendant`
```
benchmarks
    html-serializer
      deserialize
        8.52 → 12.26 ops/sec (44% faster)
      serialize
        1068.79 → 1042.88 ops/sec
    plain-serializer
      deserialize
        629.31 → 599.21 ops/sec
      serialize
        8838.64 → 7803.43 ops/sec
    rendering
      normal
        116.80 → 124.99 ops/sec
    changes
      delete-backward
        125490.67 → 131635.46 ops/sec
      delete-forward
        25364.42 → 23609.88 ops/sec
      insert-text-by-key-multiple
        94.77 → 96.53 ops/sec
      insert-text-by-key
        1033.33 → 1330.16 ops/sec
      insert-text
        1515.00 → 1263.60 ops/sec
      normalize
        2258.85 → 122505.24 ops/sec (5323% faster) 😱
      split-block
        31.98 → 70.81 ops/sec (121% faster) 🙌
    models
      from-json
        82.21 → 300.29 ops/sec (265% faster) 🙌
      get-blocks-at-range
        34000.55 → 32527.38 ops/sec
      get-blocks
        1117343.63 → 1107221.13 ops/sec
      get-characters-at-range
        16892.12 → 15218.31 ops/sec
      get-characters
        15528.78 → 15677.84 ops/sec
      get-inlines-at-range
        535680.63 → 557531.92 ops/sec
      get-inlines
        1872811.68 → 2430172.83 ops/sec
      get-leaves
        2027760.74 → 2194099.72 ops/sec
      get-marks-at-range
        2058.62 → 2257.14 ops/sec
      get-marks
        1767.86 → 1666.36 ops/sec
      get-path
        847787.16 → 991731.14 ops/sec
      get-texts-at-range
        541324.33 → 627722.89 ops/sec
      get-texts
        1077233.52 → 1241400.14 ops/sec
      has-node-multiple
        116391.48 → 135746.84 ops/sec
      has-node
        852073.43 → 1025471.67 ops/sec
      to-json
        2830.10 → 4065.79 ops/sec (44% faster)
      update-node
        13920.74 → 21126.21 ops/sec (52% faster)
```