@czechdave 2018-02-22
@ianstormtaylor @zhujinxuan should we add the benchmark from [#1650](https://github.com/ianstormtaylor/slate/issues/1650#issuecomment-367502883) that surfaced this memory issue?