@czechdave 2018-02-23
@zhujinxuan You just copy/paste the code I posted in my https://github.com/ianstormtaylor/slate/issues/1650#issuecomment-367502883 into a new file say `slate/packages/slate/benchmark/models/from-json-big.js` and that's it. That will result in an extra line in the benchmark results next to `from-json` will be `from-json-big` or whatever you name the file.

**Beware though:** as mentioned in that thread, if you run that benchmark on any version above `0.32.2` without this fix the benchmark will just crash as it runs out of memory.