@zhujinxuan 2018-02-23
@czechdave Add the benchmark by the last commit