@zhujinxuan 2018-02-23
Restore to the original logic in the commit 56023ba