@lxcid 2018-03-02
Thanks for the benchmark, which helped me a lot in identifying the root of the problem…

I want to provide my use case, hopefully it help with the PR here:

---

I have a deserialisation process that converts a custom serialised format to Slate models.

What I basically does is using `Change` to facilitate the deserialisation. The mapping is almost 1 to 1 (you can think of this serialised format as a subset of Slate models).

This serialisation usually take take sub seconds to complete.

But when the data have a lot of blocks (in my case 258 blocks), I will end up calling `Change.insertBlockAtRange()` for every block. It seems that `Change.insertBlockAtRange()` calls normalised a few times (based on console log debugging). If it matters, the block is not complex…

The outcome is that my deserialisation process now exhibits O(N^2) like behaviour. My sub seconds operation now takes 1 minutes 30 seconds to complete.

I was doing this on the server (because of Server Side Rendering) and notice memory built up when this happens. I was wondering if the memory built up is caused my memoization.

Will a sample test case help?

I might be able to write one…