@zhujinxuan 2018-03-08
@lxcid tests are very welcomed.  But I would like to notice that insertBlockAtRange has a bug about the normalization range. That is a part of reason I would not use change operation to de-serialize.